Chunk preload and no valid move preload for journey.
Preloading a few images at a time initially for journey when reaching a new node instead of preloading a lot of images immediately to make sure images to be displayed earlier in the path are loaded first. Avoiding valid moves preload when running a journey since a lot of images are unecessarily preloaded which causes black image planes because path images are loaded to slow.