Fix travis build (#639)
Summary:
Since we removed the `__init__.py` file from the `opensfm/test` folder, pytests were not running in travis. The reason was that pytest could not find the opensfm package. It is not clear why removing the `__init__.py` affected that. In any case, setting up the `PYTHONPATH` before calling pytest solves the issue.

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/639

Reviewed By: fabianschenk

Differential Revision: D23595515

Pulled By: paulinus

fbshipit-source-id: 165691f6d6714fcc8b5034db0f8629d4db6ca80a