Put all tests in a single target
Summary:
We can put all tests in a single target and still be able to call each test individually.  This should be a bit simpler to maintain.

It also makes it possible to handle all the data folder as a single target that is accessed by the single test target.  This way there is a single place (`data_generation.py`) where we check if the data is in the buck build location or in the source location.

Reviewed By: fabianschenk

Differential Revision: D23623981

fbshipit-source-id: 4421656050ff0d0b00ca33586864ae485c32986e