mapillary_sfm BUCK build
Summary:
This Diff aims at having `mapillary_sfm` building in `fbcode`, together with tests passing. Where as the `TARGETS` files are rather straightforward, having unit tests passing involved a bit more refactors :
 - All end-to-end tests (test_XXX_commands) could not be run as they were (`subporocess`).
 - Instead all code relavant to these end-to-end tests have been isolated into single function in relevant places. Now binaries are just calling these functions.
 - That involved moving a lot of code, BUT it makes it ready for `command`ification in the same spirit of `opensfm` commands.
 - Having commands paves the future for straight re-use in `fblearner`

For now, most of the end-to-end test still time-out on Sandcastle because of the `dev` mode which makes the `sfm` code super slow. We need a way to circumvent this, but `opt` mode seems to be a no-go in `fbcode`.

Reviewed By: fabianschenk, paulinus

Differential Revision: D23621252

fbshipit-source-id: 0eb2a45fecd30f93e6eb5bfa092524aa79bcf760