New features for the annotation UI
Press the 'Auto GCP' button in any of the views to visualize tracks.
The tracks are color-coded, brighter tracks are longer.
Now click on any track to automatically create a GCP from it.
Now you only have to annotate the GCP on the images that didn't get reconstructed.

Press the 'Analyze' button to get some metrics about the positional accuracy of each frame and the reprojection error of each GCP.
This information will show up in the GCP list and frame lists of each view.
It also enables the 'jump to worst GCP' function (press Q)

Press Q (after having analyzed the annotations) to jump to the GCP that has the largest average reprojection error.
Every view that has at least one annotation for this GCP will jump to the frame within the sequence that has the worst reprojection error.
Use this to find wrongly annotated points.

Smart zoom behaviour upon displaying a new frame (From scrolling or clicking on a frame list):

- If sticky zoom is set and a GCP is selected:
  + If the GCP is visible, zoom to the GCP
  + otherwise, zoom to the last clicked x, y
- If sticky zoom is not set or there is no selected GCP, the view will show all the image.

Regardless of the 'sticky zoom' mode, the view will instantly zoom in after clicking to add a GCP.

- w/s to move forwards/backwards on all sequences (linked or not)
- x as a hotkey to enable / disable sticky zoom.

A list next to each view displays a list of all the frames in the sequence.
Clicking on a row in this list immediately jumps to that frame. Each row contains:

- Frame index.
- Count of annotated points.
- Positional uncertainty of each frame. Only available if the annotations were analyzed (click 'Analyze')

So there is no wasted time arranging windows before starting the annotation.