fix: do not init glog (#653)
Summary:
We were internally initializing the glog library. This causes issues when other python libs also init it because it should only be initialized once.

This PR removes the initialization completely. If that causes issues, we can expose an init glog method to python and call it on each `__main__` on the lib.

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/653

Reviewed By: fabianschenk

Differential Revision: D24120229

Pulled By: paulinus

fbshipit-source-id: 40f41d079d34f983693e4c392204bf4914395317