Fix correspondence setup in bundle command
Summary:
The goal of this diff is to add the correspondences between the shots and the 3D points in the bundle command.

The `bundle` command was broken because it loads the reconstruction with the points and shots but without the correspondences.
Output with broken `bundle`:
{F339311050}

Note: If you can't reproduce the broken reconstruction, delete the `undistorted` folder and the `reconstruction.json` and rerun.

Reviewed By: paulinus

Differential Revision: D24218790

fbshipit-source-id: 4a7d722b02d2ed49ff97f9ab8f4a36ddf8b3d2f5