Fix alignment constraints function, syntax warning (#671)
Summary:
Hey all :wave:

I found a bug in `BAHelpers::AlignmentConstraints`. The `idx` variable is never incremented in the second for loop, causing the function to improperly populate the matrix. This in turn can affect the alignment of the reconstruction (I found at least one verified dataset that is improperly aligned without this fix).

I also fixed a warning concerning the use of `is` for comparison between strings. It's actually equivalent, but newer versions of Python complain about it.

```
$ python3 --version
Python 3.8.5
$ echo '"a" is "b"' | python3
<stdin>:1: SyntaxWarning: "is" with a literal. Did you mean "=="?
```

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/671

Reviewed By: fabianschenk, paulinus

Differential Revision: D24826315

Pulled By: YanNoun

fbshipit-source-id: cef94a04373d624258c15250311af5c52dfe00bd