feat: split same-reconstruction sequences in views
Before this commit, if --group-by-reconstruction was set,
images would be grouped attending only to the reconstruction
that they end up in.

This caused rig images to be put into the same view which was
uncomfortable for annotation.

Now, if group-by-reconstruction is set and sequence-group has
sequences in it that should be grouped (e.g. rig sequences),
we ensure that they end up in different views.