Export num views option with export_ply (#692)
Summary:
Hello :hand:

This PR proposes the addition of a new `--point-num-views` parameter to the `export_ply` command to output a new `views` containing the number of views that each point is visible from.

This lets a user generate "coverage" maps using a tool such as PDAL to identify areas that might be missing overlap.

![image](https://user-images.githubusercontent.com/1951843/104733606-f1a0bf80-570c-11eb-9c7d-a9b4a178bfe5.png)

Hope it can be useful to others.

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/692

Reviewed By: fabianschenk

Differential Revision: D26420604

Pulled By: YanNoun

fbshipit-source-id: 6516d8168c7d5255c5d0d7dc7facc3e4f18d47a0