Match undistorted image names with shot ids (#706)
Summary:
Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/706

Since we do not know on which format the input images are, we were always adding `.jpg` to the undistorted image file name before saving them. This caused a miss-match between the undistorted images' file names and the shot IDs in the undistorted reconstruction.

We fix that by renaming the shot IDs to match the image filename.

Additionally, we do not append the file format to the file name if the filename already ends with that extension. This breaks file-level backwards compatibility.

We also add the `undistorted_shot_ids.json` file to the dataset that maps original shot IDs to the list of corresponding undistorted shot IDs.  Remember that a single shot can have multiple undistorted shots.

Reviewed By: YanNoun

Differential Revision: D26443052

fbshipit-source-id: 344990760a29fbcefdb85029272a221736c04555