feat (ann. ui): Scale GPS DOP with # images
This scales the GPS DOPs so that GPS has less weight with
increasing number of images.

It is a bit of a hack to ensure that the GPS does not overwhelm the
GCP alignment cost. Usually there will be many more GPS constraints than
GCP constraints. If there is a locally consistent wrong GPS offset, they
will add up.