Make undistorted_data_path argument compulsory
Summary:
The UndistortedDataset requires a path to store the data. This was optional and `data.data_path` was used if not set. Since we do not want to assume that `data_path` exists. We make the argument required.

We add a convenience function to `opensfm.DataSet` to get the undistorted dataset without having to explicitly set that path.

Reviewed By: YanNoun

Differential Revision: D26876987

fbshipit-source-id: 7b80b835f791e47bacb72fb536fc9503f71e3a0a