feat: rig usage in SfM
Summary:
This Diff aims at actually making use of the rigs in the SfM process. To do so, we had to :
 - Modify `BAHelpers` so that they setup rig data
 - Modify `ShotNeighbohood` for `LocalBundle` so that they work with entire rig instances in case of any rig shot encountered.
 - Rewrite the `single_view` `python` adjust to N-poses adjust in C++, so that it support rigs
 - Make the `align` align instances and not shot for shot in rigs.
 - Modify the SfM loop as the following : resect a rig instance when a camera is resected (through the `add_shot` function), and triangulate point instead of assigning inliers. This is clearly the least-effort/modification way and maybe not optimal (robust triangulation ?)

For now, as we support only `hard` rig. When supporting `soft-rig` in the future, most `is_in_rig` code paths will be only for `hard-rig`.

Finally all of these is tested thanks to end-to-end test by running the whole SfM pipeline.

Reviewed By: paulinus

Differential Revision: D26845112

fbshipit-source-id: dd9b92d7744307d9f13af80c4bf79dd433b2695a