feat: rig in the bundle
Summary:
This Diff aims at refactoring the bundle so that it support rigs. To do so, we had to ;
 - We re-use the `BAData` class for expressing poses, their prior and their std dev., and their adjusted parameters.
 - We introduce a `BADataContainer` for aggregating the above data and create aggregates such as `RigInstances` (one instance pose + N rig cameras poses)
 - The above are composited under the `BANode` umbrella

This is a first step toward relying on a unified data model in the bundle. We limit the amount of new code by not refactoring other data classes, neither move code. This will done in further refactoring Diffs.

We also take the opportunity to finally have unit tests in the bundle.

In addition, we refactor the `BADataPrior`, so that a subset of parameters are constrained.

Finally, misc. missing includes were added here and there.

Reviewed By: paulinus

Differential Revision: D26845113

fbshipit-source-id: 6f6dc078429adbfe1c3789545494b5a305cdb526