Add context management for datasets
Summary:
# Context
Previously, it was impossible for the dataset to know when a command is finished. There are cases, where a something is allocated in the `__init__` and has to be cleaned up once the command is finished, e.g.
Introduce context management for all the datasets in the command runner.
# Product decisions
1. < Anything that goes beyond initial plan >
2. <  >

# Code decisions
< Anything you want to highlight so reviewers pay extra attention >

Reviewed By: YanNoun

Differential Revision: D29934062

fbshipit-source-id: 4ac1a92ff23af19c05dcd912def1856f34f23123