Annotation UI: better image pre-caching behavior + re-enable more than one view
Summary:
Previously, picking a new image that had was not cached and that had no cached neighbors would initiate a download for the whole group at the same time, slowing down the display of the image that was selected.

Now, the neighbors of the current image will only be downloaded after the current image is displayed. This makes interaction snappier.

Reviewed By: oscarlorentzon

Differential Revision: D32277665

fbshipit-source-id: ed00510ee286bdca50fd8e7a526a2fed55d878b9