fix: use map views ref counts
Summary:
`Map` data such as `shots` is exposed through view classes. We need to keep the map alive as long as some of its objects are alive in python.  For this to happen, we need
- make the views keep alive the map
- make the shots keep alive the view

A better solution will be to use pybind's own bindings for `unordered_map`. However, this is only complete in pybind 2.8 https://github.com/pybind/pybind11/pull/3310

Differential Revision: D32401524

fbshipit-source-id: 3fc30e710a2151dbb5097de8fe685ddfa80df675