explicitly add rigs to augmented shots
Summary:
When running augment, we temporarily add augmented shots to the reconstruction. When adding a shot without a rig, the `create_shot` function adds a rig instance for it. That rig instance has the same id as the shot.

That shot_id may happen to collide with an existing rig_id in the reconstruction.

To prevent this, we explicitly add temporary a temporary rig instance to the reconstruction and remove it when we are removing the shots.

Reviewed By: YanNoun

Differential Revision: D32715696

fbshipit-source-id: 4ab1ceecc0981f7e4017b2401d6fed9b253204da