fix: handling of camera in synthetic tests
Summary:
This Diff fixes the handling of cameras in synthetic data generation.

With rigs as 1st-class citizens, cameras of a given sequence should be supposed to be given as `List` and not as a single `Camera` objects.

As were not doing, so, only the first camera model given was used.

Test is modified in order to keep its behaviour.

Reviewed By: paulinus

Differential Revision: D33660484

fbshipit-source-id: 8355949cd7f11166a9f1b601b23f6d1f2a7a9799