fix: robustness to large-frame camera
Summary:
This Diff strenghten the behaviour of parrallel processing steps with images, namely feature extraction and undistortion.

We take into account the memory needed for processing a given image in both processes, in order to better pick a suitable number of // processes and queue size.

We also refactor slightly PLY export, so that no intermediate string is created and instead, data is directly flushed to file handler.

Reviewed By: mlopezantequera

Differential Revision: D34302997

fbshipit-source-id: b988b29450e2036567892492ed78489ad11af3bc