feat: add robust to common positions costs
Summary:
This Diff adds a Tukey robustifier to common position constraints.

Common position constraints historically helped with rigs that were not parametrized a such (aka. soft rigs), but can occasionnaly mess up if for some reason the device returns identical timestamps/LLA for different images, which can also happens (happended with MapSpots, happened with a Rover chunk), resulting in the scene being "shrinked" (trivial solution to reduce the error is to shrink the scene to zero)

One could also removes the common position constraint, as it use is questionnable and ad-hoc, but adding a Tukey loss makes it work.

Reviewed By: fabianschenk

Differential Revision: D35965827

fbshipit-source-id: 069f93fee4f522f310d4fd7ad0ba858bfbe2681d