max_angle argument for object triangulation
Summary:
Triangulation should have both a minimum and a maximum angle (180degrees is also not good for triangulation). This adds the max threshold.

While a good default for max_angle might be 180-min_angle, this is left as a parameter. If the keypoint location is unstable for large viewpoint changes, a max angle closer to 90 degrees or even smaller might be desirable.

Reviewed By: fabianschenk

Differential Revision: D36241730

fbshipit-source-id: 3fd874630edb8018dae78c778a07e06c415ee6b3