Remove unused-variable in mapillary/opensfm/opensfm/src/geometry/absolute_pose.h
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: fabianschenk

Differential Revision: D59586421

fbshipit-source-id: 6a8eece4ed9e6015a73021a0ee3bba4d1a0c49b6