Remove unused-variable in mapillary/opensfm/opensfm/src/bundle/error/absolute_motion_errors.h
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: meyering

Differential Revision: D64279137

fbshipit-source-id: 68f17d9881c727d38065cd80d8314d28ac325f52