Do not triangulate when the origin overlap
Summary:
When two cameras have the same origin, triangulation is ill-posed, so we skip it.

meta-only:

In f716010762, opensfm throws a lot of ceres error about residual being nan.

```
Error in evaluating the ResidualBlock.
There are two possible reasons. Either the CostFunction did not evaluate and fill all
residual and jacobians that were requested or there was a non-finite value (nan/infinite)
generated during the or jacobian computation.
Residual Block size: 4 parameter blocks x 2 residuals
For each parameter block, the value of the parameters are printed in the first column
and the value of the jacobian under the corresponding residual. If a ParameterBlock was
held constant then the corresponding jacobian is printed as 'Not Computed'. If an entry
of the Jacobian/residual array was requested but was not written to by user code, it is
indicated by 'Uninitialized'. This is an error. Residuals or Jacobian values evaluating
to Inf or NaN is also an error.
Residuals:             -nan         -nan
Parameter Block 0, size: 8

   -0.100745 | Not Computed  Not Computed
   0.0387447 | Not Computed  Not Computed
 -0.00785474 | Not Computed  Not Computed
 0.000935679 | Not Computed  Not Computed
    0.370036 | Not Computed  Not Computed
     1.00776 | Not Computed  Not Computed
 -0.00029014 | Not Computed  Not Computed
-0.000430234 | Not Computed  Not Computed
Parameter Block 1, size: 6

    -1.17687 |         -nan         -nan
    -1.15541 |         -nan         -nan
     1.17687 |         -nan         -nan
          -0 |         -nan         -nan
           0 |         -nan         -nan
           0 |         -nan         -nan
Parameter Block 2, size: 6

           0 | Not Computed  Not Computed
     -1.5708 | Not Computed  Not Computed
           0 | Not Computed  Not Computed
          -0 | Not Computed  Not Computed
           0 | Not Computed  Not Computed
           0 | Not Computed  Not Computed
Parameter Block 3, size: 3

           0 |         -nan         -nan
           0 |         -nan         -nan
           0 |         -nan         -nan
```

This is because the data used in this experiment is in rigs. Each rig has two cameras, facing back to back, sharing identical origin (camera center).

When there is a track that has observations in both images in the same rig, there is a problem. The point will be triangulated at [0,0,0] because that's the only place that can fulfill the two bearing vectors. Currently that is labeled as a valid solution.

With the ceres errors, BAs are effectively not run and the map is simply an unoptimized result of the resections

{F1976773050}

This diff avoids triangulating when the camera centers are the same point, thus avoiding the ceres nan errors.

Differential Revision: D72404151

fbshipit-source-id: 856f2cda3c92cbeda29085a991d37947fd3b2663