Remove max_angle parameter
Summary:
We had an extra max_angle parameter in the midpoint triangulation function. The passed value was always `pi-min_angle`. The DLT triangulation function did not check for max angle.

So we remove the parameter, use always `pi-min_angle` internally and add the check also in the DLT triangulation function.

Reviewed By: YanNoun

Differential Revision: D72560210

fbshipit-source-id: bd355fd5a6748f4010ab6077fd6ac28a2171325e