reduce 70% memory usage during reconstructability computation
Summary:
In order to compute the list of image pairs (sorted by reconstructability), the existing code load a giant dictionary of <im1, im2> -> <feature_array_im1, feature_array_im2, track ids> in memory. It's expensive to save per-image features in memory, and even more expensive to save the feature array repeatedly when an image is matched with many images.
This change adds a new version of the image pair computation logic. It avoids loading all the feature arrays in memory, but loads the feature arrays lazily when the corresponding images are touched. As a result, this can save more than 70% of memory of that piece of code, which has been a memory bottleneck for opensfm for a long time.

meta-only:

Reviewed By: DodgySpaniard

Differential Revision: D78921836

fbshipit-source-id: 94c7385970b1fde1bcf5c056d70a9dfa4e43364f