expose aspect_ratio_sd parameter
Summary:
It is questionable to assign the same std for focal length and aspect ratio, because aspect ratio, being the ratio of fu and fv, should be one order of magnitude smaller than fu and fv themselves.

In practice we have noticed problematic maps built from opensfm BA that have very wrong aspect ratio, causing the camera poses to be inaccurate.

Reviewed By: paulinus

Differential Revision: D78990090

fbshipit-source-id: 65334567a693ab8872e91746942ef011f2e62de6