fix: default to numpy.core if available (#1113)
Summary:
Not sure how to best handle access to `numpy.core` (in numpy 1) and `numpy._core` (in numpy 2). It seems the best would be not to access it at all.

In any case, there are situations in numpy 1 where `numpy._core` exists but does not contain any `multiarray` definition. So we default to use `numpy.core` if it exists and only use `numpy._core` otherwise.

Tested locally using `uv`. Failed before this fix.

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/1113

Differential Revision: D86282715

Pulled By: paulinus

fbshipit-source-id: eca406c318237fb18db19dd27611ddfd903eb1ef